Show dynamic legend on execute key
Count parentheses levels and use that count to display `)` on the
execute key when it will pop a parenthesis.

In this version an expression that starts with `(` is a bit weird:
poppin that parenthesis will take you all the way back to `=` result
mode, not `..) =` result mode. This happens because those two modes are
still the same. I'm not sure yet if I'll split them into two.